photorag_QA/requirements.txt
rich>=13.0.0
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.0.0
click>=8.0.0
tenacity>=8.0.0
tiktoken>=0.5.0
